requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
openpyxl
gensim
spacy
transformers
altair
sentencepiece
torch
huggingface_hub
openai